requirements.txt
streamlit==1.22.0
pandas==1.5.3
plotly==5.14.1
requests==2.28.2
numpy==1.24.3
kafka-python==2.0.2
hdfs==2.7.0
pyarrow==11.0.0
python-snappy==0.7.1
